pyproject.toml
[project]
name = "ai-companion-saas"
version = "1.0.0"
description = "AI Companion SAAS — personalised virtual partner experiences powered by Gemini AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.65.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-community>=0.3.0",
    "duckduckgo-search>=6.0.0",
    "upstash-redis>=1.1.0",
    "langchain-openai>=0.2.0",
    "langchain-groq>=0.1.0",
    "tavily-python>=0.1.0",
    "sqlalchemy>=2.0.35",
    "psycopg[binary]>=3.2.3",
    "python-jose[cryptography]>=3.3.0",
]
